+++ R/quilting_hut.R
@@ -32,6 +32,9 @@ create_width_balanced_cover <- function(min_val,
                                         num_bins,
                                         percent_overlap) {
 
+  if (any(is.na(c(min_val, max_val, num_bins, percent_overlap)))) {
+    stop("cannot create a cover with an NA input!")
+  }
   if ((num_bins <= 0) | (!(num_bins %% 1 == 0))) {
     stop("number of bins must be a positive integer")
   } else if ((percent_overlap < 0) | (percent_overlap > 100)) {