+++ NEWS.md
@@ -1,3 +1,8 @@
+# mappeR 1.1.0
+
+* Added safety checks for mapper input functions (no `NA` inputs, etc)
+* Removed Cytoscape interfacing for the release version; will remain in a separate branch
+
 # mappeR 1.0.0
 
 * Balls/bins with equal amounts of data no longer cause wonkiness